Rename cosmicconnector to (Simple)InterconnectProvider
