+++ examples/golang/simpleinterconnectprovider/README.md
