+++ examples/golang/simpleinterconnectprovider/config/BUILD.bazel
@@ -29,7 +29,7 @@ proto_library(
 
 go_proto_library(
     name = "config_go_proto",
-    importpath = "github.com/outernetcouncil/federation/gen/go/examples/golang/cosmicconnector/config",
+    importpath = "github.com/outernetcouncil/federation/gen/go/examples/golang/simpleinterconnectprovider/config",
     proto = ":config_proto",
 )
 
@@ -39,7 +39,7 @@ go_library(
         "config.go",
         "doc.go",
     ],
-    importpath = "github.com/outernetcouncil/federation/examples/golang/cosmicconnector/config",
+    importpath = "github.com/outernetcouncil/federation/examples/golang/simpleinterconnectprovider/config",
     deps = [
         ":config_go_proto",
         "@com_github_rs_zerolog//:zerolog",
