+++ examples/golang/simpleinterconnectprovider/config/doc.go
